Update jquery.vmap.js
Added an option for $('#map').vectorMap('close') (assuming $('#map') has already been instantiated). 
This helped to fix the issue of window resize events being added with each new map but not removed when the map is removed.  When a new initialization call is made to .vectorMap() a name-spaced resize event specific to that instantiation is created and placed in a registry.  One window resize event handler is declared and it triggers all the name-spaced events within the registry (limited to once every 10 ms for performance reasons).  When .vectorMap('close') is called it removes the name-spaced event from the registry, unbinds event listeners on the svg element, callbacks, etc and removes the SVG from the DOM.

Also updated the tooltip functionality so that the options object can now take tooltipX and tooltipY attributes of 'right' and 'bottom' (respectively) which will position the tooltip to the right or bottom of the cursor.  These attributes default to 'left' and 'top' respectively to mimic the prior default behavior.  The spacing now also accounts for dynamic padding of the tooltip element.